Factor out Options into its own class - getting crowed inside Config
