+++ EDDiscovery/EDDiscovery.csproj
@@ -192,6 +192,7 @@
     <Compile Include="Actions\ActionsEDDCmds\ActionMenuItem.cs" />
     <Compile Include="Actions\ActionVars.cs" />
     <Compile Include="Actions\ConditionFunctionsEDD.cs" />
+    <Compile Include="EDDOptions.cs" />
     <Compile Include="GridControl.cs">
       <SubType>UserControl</SubType>
     </Compile>
